+++ next/pages/guidelines/typography.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import type { InferGetStaticPropsType } from 'next';
 import { ContentActionsCheckSmall } from '@thumbtack/thumbprint-icons';
-import { ContentPage } from '../../../components/mdx/mdx';
-import getContentPageStaticProps from '../../../utils/get-content-page-static-props';
-import { Code, H2, P, Table, TD, TH } from '../../../components/mdx/components';
+import { ContentPage } from '../../components/mdx/mdx';
+import getContentPageStaticProps from '../../utils/get-content-page-static-props';
+import { Code, H2, P, Table, TD, TH } from '../../components/mdx/components';
 
-import Alert from '../../../components/alert/alert';
-import InlineCode from '../../../components/inline-code/inline-code';
+import Alert from '../../components/alert/alert';
+import InlineCode from '../../components/inline-code/inline-code';
 
 export const getStaticProps = getContentPageStaticProps;
