+++ next/utils/get-layout-props.ts
@@ -57,21 +57,21 @@ export default function getLayoutProps(): LayoutProps {
             },
             {
                 title: 'Guidelines',
-                href: '/guide/product/aspect-ratio',
+                href: '/guidelines/aspect-ratio',
                 groups: [
                     [
-                        { title: 'Aspect ratio', href: '/guide/product/aspect-ratio' },
-                        { title: 'Brand assets', href: '/guide/product/brand-assets' },
-                        { title: 'Breakpoints', href: '/guide/product/breakpoints' },
-                        { title: 'Color', href: '/guide/product/color' },
-                        { title: 'Design tokens', href: '/guide/product/design-tokens' },
-                        { title: 'Iconography', href: '/guide/product/iconography' },
-                        { title: 'Loaders', href: '/guide/product/loaders' },
-                        { title: 'Motion', href: '/guide/product/motion' },
-                        { title: 'Spacers', href: '/guide/product/spacers' },
-                        { title: 'Toolkits', href: '/guide/product/toolkits' },
-                        { title: 'Truncation', href: '/guide/product/truncation' },
-                        { title: 'Typography', href: '/guide/product/typography' },
+                        { title: 'Aspect ratio', href: '/guidelines/aspect-ratio' },
+                        { title: 'Brand assets', href: '/guidelines/brand-assets' },
+                        { title: 'Breakpoints', href: '/guidelines/breakpoints' },
+                        { title: 'Color', href: '/guidelines/color' },
+                        { title: 'Design tokens', href: '/guidelines/design-tokens' },
+                        { title: 'Iconography', href: '/guidelines/iconography' },
+                        { title: 'Loaders', href: '/guidelines/loaders' },
+                        { title: 'Motion', href: '/guidelines/motion' },
+                        { title: 'Spacers', href: '/guidelines/spacers' },
+                        { title: 'Toolkits', href: '/guidelines/toolkits' },
+                        { title: 'Truncation', href: '/guidelines/truncation' },
+                        { title: 'Typography', href: '/guidelines/typography' },
                     ],
                 ],
             },
